# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateSales:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      payment_type:
        type: string
      price:
        type: integer
      shop_assistent_id:
        type: string
      status:
        type: string
    type: object
...
  models.UpdateSales:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      id:
        type: string
      payment_type:
        type: string
      price:
        type: integer
      shop_assistent_id:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /sales:
    get:
      consumes:
      - application/json
      description: Get List Sales
      operationId: get_list_sales
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Sales
      tags:
      - Sales
    post:
      consumes:
      - application/json
      description: Create Sales
      operationId: create_sales
      parameters:
      - description: CreateSalesRequest
        in: body
        name: Sales
        required: true
        schema:
          $ref: '#/definitions/models.CreateSales'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Sales
      tags:
      - Sales
  /sales/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Sales
      operationId: delete_sales
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Sales
      tags:
      - Sales
    get:
      consumes:
      - application/json
      description: Get By ID Sales
      operationId: get_by_id_sales
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Sales
      tags:
      - Sales
    put:
      consumes:
      - application/json
      description: Update Sales
      operationId: update_sales
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateSalesRequest
        in: body
        name: Sales
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSales'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Sales
      tags:
      - Sales
...
swagger: "2.0"